refactor(node): tests for linked_list_node
